+++ medicines/doc-index-updater/src/state_manager/mod.rs
@@ -5,6 +5,7 @@ use crate::{
     models::{JobStatus, JobStatusResponse, XMLJobStatusResponse},
 };
 use ::redis::Client;
+use async_trait::async_trait;
 use uuid::Uuid;
 use warp::{http::StatusCode, reply::Json, Filter, Rejection, Reply};
 mod redis;
@@ -14,7 +15,7 @@ pub struct StateManager {
     pub client: Client,
 }
 
-#[async_trait::async_trait]
+#[async_trait]
 pub trait JobStatusClient {
     async fn get_status(&self, id: Uuid) -> Result<JobStatusResponse, MyRedisError>;
     async fn set_status(
@@ -22,6 +23,27 @@ pub trait JobStatusClient {
         id: Uuid,
         status: JobStatus,
     ) -> Result<JobStatusResponse, MyRedisError>;
+}
+
+#[cfg(test)]
+pub struct TestJobStatusClient {}
+
+#[cfg(test)]
+#[async_trait]
+impl JobStatusClient for TestJobStatusClient {
+    async fn get_status(
+        &self,
+        _id: Uuid,
+    ) -> Result<crate::models::JobStatusResponse, crate::state_manager::MyRedisError> {
+        unimplemented!()
+    }
+    async fn set_status(
+        &self,
+        id: Uuid,
+        status: JobStatus,
+    ) -> Result<crate::models::JobStatusResponse, crate::state_manager::MyRedisError> {
+        Ok(JobStatusResponse { id, status })
+    }
 }
 
 impl StateManager {
